Added a link to forecast discussions. Data files are now encoded with floats rather than doubles. The application icon now displays the current conditions icon as a rollover image.

+++ src/java/Station.java
@@ -150,8 +150,8 @@ public abstract class Station
 			}
 		}
 		sta.name = new String(buff);
-		sta.lat_ = din.readDouble();
-		sta.lon_ = din.readDouble();
+		sta.lat_ = (double)din.readFloat();
+		sta.lon_ = (double)din.readFloat();
 		
 		byte[] buff2 = new byte[1]; // goUp code
 		bytes = din.read(buff2);
